backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "enka>=2.5.8",
    "fastapi-restful[all]>=0.6.0",
    "fastapi-utils>=0.8.0",
    "fastapi[standard]>=0.116.0",
    "hakushin-py",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.35.0",
    "yatta-py>=1.3.12",
]

[tool.uv.sources]
hakushin-py = { git = "https://github.com/seriaati/hakushin-py.git" }
